+++ jdkjson-api/src/main/java/org/jdkstack/jdkjson/api/reader/verson1/JsonReader.java
@@ -8,19 +8,5 @@ package org.jdkstack.jdkjson.api.reader.verson1;
  * @author admin
  */
 public interface JsonReader {
-
-  /**
-   * 处理冒号.
-   *
-   * @author admin
-   */
-  void colon();
-
-  /**
-   * .
-   *
-   * @return String String.
-   * @author admin
-   */
-  String comment();
+  //
 }
