+++ jdkjson-core/src/main/java/org/jdkstack/jdkjson/core/reader/value2/BaseValue.java
@@ -18,6 +18,7 @@ public class BaseValue implements Value {
 
   @Override
   public Object deserialisation() {
+
     return null;
   }
 }
